requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.34
python-dotenv==0.21.0